requirements.txt
cognitive_face
azure-keyvault
azure-mgmt-nspkg
azure-mgmt-resource
azure-nspkg
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
cryptography==2.2.2
entrypoints==0.2.3
Flask==0.12.1
Flask-PyMongo==2.0.0
Flask-Script==2.0.5
flask-mongo-sessions
flask_mongoengine
keyring==13.2.1
msrest==0.5.1
msrestazure==0.4.34
oauthlib==2.1.0
packaging==16.8
psycopg2==2.7.1
pycparser==2.18
PyJWT==1.6.4
pymongo==3.5.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-editor==1.0.3
pywin32-ctypes==0.1.2
requests==2.19.1
requests-oauthlib==1.0.0
six==1.10.0
SQLAlchemy==1.1.9
typing==3.6.4
urllib3==1.23
waitress==1.0.1
Werkzeug==0.12.1
http==0.2
Pillow